requirements.txt
yagmail~=0.15.293
fpdf~=1.7.2
Flask~=3.1.0
WTForms~=3.2.1
requests~=2.32.3
pillow~=11.0.0
python-dotenv~=1.0.1